requirements.txt
numpy
opencv-python
Pillow
torch>=2.0.0
torchvision
tqdm
ultralytics 